+++ example/lib/main.dart
@@ -23,24 +23,28 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   String _fairbidSdkVersion = 'Unknown';
 
-  FairBid _sdk;
+  FairBid? _sdk;
 
   bool _enableLogs = true;
 
   bool _muteAds = false;
 
-  TextEditingController _sdkIdController;
+  late TextEditingController _sdkIdController;
 
-  String _appId;
+  String? _appId;
 
-  int _step;
+  int _step = 0;
 
-  Stream<ImpressionData> _impressionStream;
+  late StreamSubscription<MediationAdapterStartEvent> _adapterStartedEventsSub;
 
-  Stream<ImpressionData> get impressionsStream {
+  bool get _sdkAvailable => _sdk != null;
+
+  Stream<ImpressionData?>? _impressionStream;
+
+  Stream<ImpressionData?>? get impressionsStream {
     if (_impressionStream == null) {
-      _impressionStream = Stream.fromFuture(_sdk.started)
-          .asyncExpand((_) => _sdk.events
+      _impressionStream = Stream.fromFuture(_sdk!.started)
+          .asyncExpand((_) => _sdk!.events
               .where((event) => event.impressionData != null)
               .map((event) => event.impressionData))
           .asBroadcastStream();
@@ -53,12 +57,12 @@ class _MyAppState extends State<MyApp> {
     super.initState();
     initPlatformState();
     _sdkIdController = TextEditingController();
-    _step = 0;
   }
 
   @override
   void dispose() {
     _sdkIdController.dispose();
+    _adapterStartedEventsSub.cancel();
     super.dispose();
   }
 
@@ -99,26 +103,26 @@ class _MyAppState extends State<MyApp> {
                   currentStep: _step,
                   onStepTapped: (step) => setState(() => _step = step),
                   controlsBuilder: (context,
-                          {VoidCallback onStepContinue,
-                          VoidCallback onStepCancel}) =>
+                          {VoidCallback? onStepContinue,
+                          VoidCallback? onStepCancel}) =>
                       Container(),
                   steps: <Step>[
                     Step(
                       title: Text("Setup SDK"),
-                      isActive: _sdk == null,
+                      isActive: !_sdkAvailable,
                       subtitle: _appId == null ? null : Text("App ID: $_appId"),
                       content: _buildFirstStep(),
                     ),
                     Step(
                       title: Text("User data"),
-                      isActive: _sdk != null,
+                      isActive: _sdkAvailable,
                       subtitle: Text("Optional"),
-                      content: _sdk != null ? UserDataForm() : Container(),
+                      content: _sdkAvailable ? UserDataForm() : Container(),
                     ),
                     Step(
-                      isActive: _sdk != null,
+                      isActive: _sdkAvailable,
                       title: Text("Ads control"),
-                      subtitle: _sdk != null
+                      subtitle: _sdkAvailable
                           ? ImpressionPresenter(
                               [AdType.interstitial, AdType.rewarded],
                               impressionsStream,
@@ -127,34 +131,35 @@ class _MyAppState extends State<MyApp> {
                                 RewardedAd.impressionDepth
                               ])
                           : null,
-                      content: _sdk != null
+                      content: _sdkAvailable
                           ? _buildSdkWidgets(context)
                           : Container(),
                     ),
                     Step(
-                      isActive: _sdk != null,
+                      isActive: _sdkAvailable,
                       title: Text("Banners control"),
-                      subtitle: _sdk != null
+                      subtitle: _sdkAvailable
                           ? ImpressionPresenter([AdType.banner],
                               impressionsStream, [BannerAd.impressionDepth])
                           : null,
                       content:
-                          _sdk != null ? BannerAds(sdk: _sdk) : Container(),
+                          _sdkAvailable ? BannerAds(sdk: _sdk!) : Container(),
                     ),
                     Step(
-                      isActive: _sdk != null,
+                      isActive: _sdkAvailable,
                       title: Text("Banner views control"),
                       subtitle: Text(
                         "Experimental",
                         style: TextStyle(color: Colors.deepOrangeAccent),
                       ),
-                      content:
-                          _sdk != null ? BannerViewAds(sdk: _sdk) : Container(),
+                      content: _sdkAvailable
+                          ? BannerViewAds(sdk: _sdk!)
+                          : Container(),
                     ),
                     Step(
-                      isActive: _sdk != null,
+                      isActive: _sdkAvailable,
                       title: Text("Test suite"),
-                      content: OutlineButton(
+                      content: OutlinedButton(
                         child: Text("Open Test suite"),
                         onPressed: () => _sdk?.showTestSuite(),
                       ),
@@ -193,7 +198,7 @@ class _MyAppState extends State<MyApp> {
               ),
               Padding(
                 padding: EdgeInsets.symmetric(horizontal: 6.0),
-                child: OutlineButton(
+                child: OutlinedButton(
                   onPressed: () {
                     _initSDK();
                   },
@@ -221,9 +226,9 @@ class _MyAppState extends State<MyApp> {
 
   Widget _buildSdkWidgets(BuildContext context) {
     return FutureBuilder<bool>(
-      future: _sdk.started,
-      builder: (context, snapshot) => snapshot.hasData && snapshot.data
-          ? FullScreenAds(sdk: _sdk)
+      future: _sdk!.started,
+      builder: (context, snapshot) => snapshot.hasData && snapshot.data!
+          ? FullScreenAds(sdk: _sdk!)
           : Container(
               child: Center(
                 child: CircularProgressIndicator(),
@@ -238,6 +243,13 @@ class _MyAppState extends State<MyApp> {
       debugLogging: _enableLogs,
       loggingLevel: _enableLogs ? LoggingLevel.info : LoggingLevel.error,
     ));
+    _adapterStartedEventsSub = sdk.adapterEventsStream.listen((event) {
+      print(
+          'network ${event.networkName}(${event.networkVersion}) has ${event.successful ? '' : 'not '}started');
+      if (event.errorMessage != null) {
+        print('error: ${event.errorMessage}');
+      }
+    });
     await sdk.started;
     setState(() {
       _sdk = sdk;
@@ -250,12 +262,12 @@ class _MyAppState extends State<MyApp> {
 
 class ImpressionPresenter extends StatelessWidget {
   final List<AdType> adTypes;
-  final Stream<ImpressionData> impressions;
-  final List<Future<int>> initialImpressions;
+  final Stream<ImpressionData?>? impressions;
+  final List<Future<int?>> initialImpressions;
 
   const ImpressionPresenter(
       this.adTypes, this.impressions, this.initialImpressions,
-      {Key key})
+      {Key? key})
       : super(key: key);
 
   @override
@@ -263,15 +275,15 @@ class ImpressionPresenter extends StatelessWidget {
     return Row(
       children: <Widget>[
         for (var index = 0; index < adTypes.length; index++)
-          StreamBuilder<int>(
+          StreamBuilder<int?>(
             initialData: 0,
             stream: Rx.concat([
               initialImpressions[index].asStream(),
-              impressions
-                  .where((imp) => imp.placementType == adTypes[index])
-                  .map((imp) => imp.impressionDepth)
+              impressions!
+                  .where((imp) => imp!.placementType == adTypes[index])
+                  .map((imp) => imp!.impressionDepth)
             ]),
-            builder: (BuildContext context, AsyncSnapshot<int> snapshot) {
+            builder: (BuildContext context, AsyncSnapshot<int?> snapshot) {
               return Chip(label: Text('${snapshot.data}'));
             },
           ),