+++ lib/src/privacy.dart
@@ -9,16 +9,16 @@ class PrivacySettings {
   /// The [consentString] should be a correct string formatted according to [IAB specification](https://github.com/InteractiveAdvertisingBureau/GDPR-Transparency-and-Consent-Framework/blob/master/Consent%20string%20and%20vendor%20list%20formats%20v1.1%20Final.md)
   ///
   static Future<void> updateGDPRConsent(
-      {@required bool grantsConsent, String consentString}) {
-    assert(grantsConsent != null);
+      {bool? grantsConsent, String? consentString}) {
     final params = <String, Object>{
-      "grantConsent": grantsConsent,
+      if (grantsConsent != null) "grantConsent": grantsConsent,
       if (consentString != null) "consentString": consentString,
     };
     return FairBidInternal._channel.invokeMethod("updateGDPR", params);
   }
 
   /// Clears GDPR related data
+  @Deprecated('Use `updateGDPRConsent(consentString: null)` instead')
   static Future<void> clearGDPRConsent() =>
       FairBidInternal._channel.invokeMethod("clearGDPR");
 
@@ -26,7 +26,7 @@ class PrivacySettings {
   ///
   /// The `consentString` should be a correct string formatted according to [IAB specification](https://iabtechlab.com/standards/ccpa/)
   ///
-  static Future<void> updateCCPAString({String ccpaString}) {
+  static Future<void> updateCCPAString({String? ccpaString}) {
     final params = <String, Object>{
       if (ccpaString != null) "ccpaString": ccpaString,
     };