+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### 1.1.0
+* Updated dependencies to **FairBid 3.18.0** (Official changelog: [Android](https://developer.fyber.com/hc/en-us/articles/360010205178-FairBid-Android-SDK-Changelog), [iOS](https://developer.fyber.com/hc/en-us/articles/360010021878-FairBid-iOS-SDK-Changelog)).
+* **New API:** [FairBid.adapterEventsStream] - stream of [MediationAdapterStartEvent]s to track which adapters has been started. (Official documentation: [Android](https://developer.fyber.com/hc/en-us/articles/4405766010641-Adapter-Started-Callback), [iOS](https://developer.fyber.com/hc/en-us/articles/4405738348305-Adapter-Started-Callback)).
+* Fixed issue in `BannerView` - the banner should be rendered in the view instead of popping out at the bottom of the screen.
+
+### 1.0.1
+* Updated dependencies to **FairBid 3.14.0** (Official changelog: [Android](https://developer.fyber.com/hc/en-us/articles/360010205178-FairBid-Android-SDK-Changelog), [iOS](https://developer.fyber.com/hc/en-us/articles/360010021878-FairBid-iOS-SDK-Changelog)).
+  * Deprecated `PrivacySettings.clearGDPRConsent()`.
+
+### 1.0.0
+* **Breaking change:** `Null-safety` support.
+* Updated dependencies to **FairBid 3.13.0** (Official changelog: [Android](https://developer.fyber.com/hc/en-us/articles/360010205178-FairBid-Android-SDK-Changelog), [iOS](https://developer.fyber.com/hc/en-us/articles/360010021878-FairBid-iOS-SDK-Changelog)).
+
 ### 0.11.2
 * Updated dependencies to **FairBid 3.12.0** (Official changelog: [Android](https://developer.fyber.com/hc/en-us/articles/360010205178-FairBid-Android-SDK-Changelog), [iOS](https://developer.fyber.com/hc/en-us/articles/360010021878-FairBid-iOS-SDK-Changelog))
 * Fixed [crash on iOS](https://github.com/ukasz123/fairbid_flutter/pull/21) when attempting to show full screen ads.